requirements.txt
matplotlib==3.1.3
numpy==1.21.1
pandas==1.0.1
Pillow==9.0.1
scikit_learn==1.0.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.3
